+++ src/imap/model.rs
@@ -1,20 +1,10 @@
-use error_chain::error_chain;
+use anyhow::{anyhow, Context, Result};
 use imap;
 use log::{debug, trace};
 use native_tls::{self, TlsConnector, TlsStream};
 use std::{collections::HashSet, convert::TryFrom, iter::FromIterator, net::TcpStream};
 
-use crate::config::model::Account;
-use crate::ctx::Ctx;
-use crate::flag::model::Flags;
-use crate::msg::model::Msg;
-
-error_chain! {
-    links {
-        Config(crate::config::model::Error, crate::config::model::ErrorKind);
-        MessageError(crate::msg::model::Error, crate::msg::model::ErrorKind);
-    }
-}
+use crate::{config::model::Account, ctx::Ctx, flag::model::Flags, msg::model::Msg};
 
 /// A little helper function to create a similiar error output. (to avoid duplicated code)
 fn format_err_msg(description: &str, account: &Account) -> String {
@@ -56,7 +46,7 @@ impl<'a> ImapConnector<'a> {
             .danger_accept_invalid_certs(insecure)
             .danger_accept_invalid_hostnames(insecure)
             .build()
-            .chain_err(|| format_err_msg("Could not create TLS connector", account))?;
+            .context(format_err_msg("cannot create TLS connector", account))?;
 
         debug!("create client");
         let mut client_builder = imap::ClientBuilder::new(&account.imap_host, account.imap_port);
@@ -66,13 +56,13 @@ impl<'a> ImapConnector<'a> {
         }
         let client = client_builder
             .connect(|domain, tcp| Ok(TlsConnector::connect(&ssl_conn, domain, tcp)?))
-            .chain_err(|| format_err_msg("Could not connect to IMAP server", account))?;
+            .context(format_err_msg("cannot connect to IMAP server", account))?;
 
         debug!("create session");
         let sess = client
             .login(&account.imap_login, &account.imap_passwd()?)
             .map_err(|res| res.0)
-            .chain_err(|| format_err_msg("Could not login to IMAP server", account))?;
+            .context(format_err_msg("cannot login to IMAP server", account))?;
 
         Ok(Self { account, sess })
     }
@@ -113,11 +103,11 @@ impl<'a> ImapConnector<'a> {
 
         self.sess
             .select(mbox)
-            .chain_err(|| format!("Could not select mailbox `{}`", mbox))?;
+            .context(format!("cannot select mailbox `{}`", mbox))?;
 
         self.sess
             .uid_store(uid_seq, format!("FLAGS ({})", flags))
-            .chain_err(|| format!("Could not set flags `{}`", &flags))?;
+            .context(format!("cannot set flags `{}`", &flags))?;
 
         Ok(())
     }
@@ -147,11 +137,11 @@ impl<'a> ImapConnector<'a> {
 
         self.sess
             .select(mbox)
-            .chain_err(|| format!("Could not select mailbox `{}`", mbox))?;
+            .context(format!("cannot select mailbox `{}`", mbox))?;
 
         self.sess
             .uid_store(uid_seq, format!("+FLAGS ({})", flags))
-            .chain_err(|| format!("Could not add flags `{}`", &flags))?;
+            .context(format!("cannot add flags `{}`", &flags))?;
 
         Ok(())
     }
@@ -163,11 +153,11 @@ impl<'a> ImapConnector<'a> {
 
         self.sess
             .select(mbox)
-            .chain_err(|| format!("Could not select mailbox `{}`", mbox))?;
+            .context(format!("cannot select mailbox `{}`", mbox))?;
 
         self.sess
             .uid_store(uid_seq, format!("-FLAGS ({})", flags))
-            .chain_err(|| format!("Could not remove flags `{}`", &flags))?;
+            .context(format!("cannot remove flags `{}`", &flags))?;
 
         Ok(())
     }
@@ -176,7 +166,7 @@ impl<'a> ImapConnector<'a> {
         let uids: Vec<u32> = self
             .sess
             .uid_search("NEW")
-            .chain_err(|| "Could not search new messages")?
+            .context("cannot search new messages")?
             .into_iter()
             .collect();
         debug!("found {} new messages", uids.len());
@@ -189,7 +179,7 @@ impl<'a> ImapConnector<'a> {
         debug!("examine mailbox: {}", &ctx.mbox);
         self.sess
             .examine(&ctx.mbox)
-            .chain_err(|| format!("Could not examine mailbox `{}`", &ctx.mbox))?;
+            .context(format!("cannot examine mailbox `{}`", &ctx.mbox))?;
 
         debug!("init messages hashset");
         let mut msgs_set: HashSet<u32> =
@@ -208,7 +198,7 @@ impl<'a> ImapConnector<'a> {
                         false
                     })
                 })
-                .chain_err(|| "Could not start the idle mode")?;
+                .context("cannot start the idle mode")?;
 
             let uids: Vec<u32> = self
                 .search_new_msgs()?
@@ -227,12 +217,12 @@ impl<'a> ImapConnector<'a> {
                 let fetches = self
                     .sess
                     .uid_fetch(uids, "(ENVELOPE)")
-                    .chain_err(|| "Could not fetch new messages enveloppe")?;
+                    .context("cannot fetch new messages enveloppe")?;
 
                 for fetch in fetches.iter() {
                     let msg = Msg::try_from(fetch)?;
                     let uid = fetch.uid.ok_or_else(|| {
-                        format!("Could not retrieve message {}'s UID", fetch.message)
+                        anyhow!(format!("cannot retrieve message {}'s UID", fetch.message))
                     })?;
 
                     let subject = msg.headers.subject.clone().unwrap_or_default();
@@ -255,7 +245,7 @@ impl<'a> ImapConnector<'a> {
         debug!("examine mailbox: {}", &ctx.mbox);
         self.sess
             .examine(&ctx.mbox)
-            .chain_err(|| format!("Could not examine mailbox `{}`", &ctx.mbox))?;
+            .context(format!("cannot examine mailbox `{}`", &ctx.mbox))?;
 
         loop {
             debug!("begin loop");
@@ -269,7 +259,7 @@ impl<'a> ImapConnector<'a> {
                         false
                     })
                 })
-                .chain_err(|| "Could not start the idle mode")?;
+                .context("cannot start the idle mode")?;
             ctx.config.exec_watch_cmds(&ctx.account)?;
             debug!("end loop");
         }
@@ -279,7 +269,7 @@ impl<'a> ImapConnector<'a> {
         let names = self
             .sess
             .list(Some(""), Some("*"))
-            .chain_err(|| "Could not list mailboxes")?;
+            .context("cannot list mailboxes")?;
 
         Ok(names)
     }
@@ -293,7 +283,7 @@ impl<'a> ImapConnector<'a> {
         let last_seq = self
             .sess
             .select(mbox)
-            .chain_err(|| format!("Could not select mailbox `{}`", mbox))?
+            .context(format!("cannot select mailbox `{}`", mbox))?
             .exists as i64;
 
         if last_seq == 0 {
@@ -313,7 +303,7 @@ impl<'a> ImapConnector<'a> {
         let fetches = self
             .sess
             .fetch(range, "(UID FLAGS ENVELOPE INTERNALDATE)")
-            .chain_err(|| "Could not fetch messages")?;
+            .context("cannot fetch messages")?;
 
         Ok(Some(fetches))
     }
@@ -327,14 +317,17 @@ impl<'a> ImapConnector<'a> {
     ) -> Result<Option<imap::types::ZeroCopy<Vec<imap::types::Fetch>>>> {
         self.sess
             .select(mbox)
-            .chain_err(|| format!("Could not select mailbox `{}`", mbox))?;
+            .context(format!("cannot select mailbox `{}`", mbox))?;
 
         let begin = page * page_size;
         let end = begin + (page_size - 1);
         let uids: Vec<String> = self
             .sess
             .search(query)
-            .chain_err(|| format!("Could not search in `{}` with query `{}`", mbox, query))?
+            .context(format!(
+                "cannot search in `{}` with query `{}`",
+                mbox, query
+            ))?
             .iter()
             .map(|seq| seq.to_string())
             .collect();
@@ -347,7 +340,7 @@ impl<'a> ImapConnector<'a> {
         let fetches = self
             .sess
             .fetch(&range, "(UID FLAGS ENVELOPE INTERNALDATE)")
-            .chain_err(|| format!("Could not fetch range `{}`", &range))?;
+            .context(format!("cannot fetch range `{}`", &range))?;
 
         Ok(Some(fetches))
     }
@@ -356,15 +349,15 @@ impl<'a> ImapConnector<'a> {
     pub fn get_msg(&mut self, mbox: &str, uid: &str) -> Result<Msg> {
         self.sess
             .select(mbox)
-            .chain_err(|| format!("Could not select mailbox `{}`", mbox))?;
+            .context(format!("cannot select mailbox `{}`", mbox))?;
 
         match self
             .sess
             .uid_fetch(uid, "(FLAGS BODY[] ENVELOPE INTERNALDATE)")
-            .chain_err(|| "Could not fetch bodies")?
+            .context("cannot fetch bodies")?
             .first()
         {
-            None => Err(format!("Could not find message `{}`", uid).into()),
+            None => Err(anyhow!(format!("cannot find message `{}`", uid))),
             Some(fetch) => Ok(Msg::try_from(fetch)?),
         }
     }
@@ -378,7 +371,7 @@ impl<'a> ImapConnector<'a> {
             .append(mbox, &body)
             .flags(flags)
             .finish()
-            .chain_err(|| format!("Could not append message to `{}`", mbox))?;
+            .context(format!("cannot append message to `{}`", mbox))?;
 
         Ok(())
     }
@@ -386,7 +379,7 @@ impl<'a> ImapConnector<'a> {
     pub fn expunge(&mut self, mbox: &str) -> Result<()> {
         self.sess
             .expunge()
-            .chain_err(|| format!("Could not expunge `{}`", mbox))?;
+            .context(format!("cannot expunge `{}`", mbox))?;
 
         Ok(())
     }