+++ src/smtp.rs
@@ -1,4 +1,4 @@
-use error_chain::error_chain;
+use anyhow::Result;
 use lettre::{
     self,
     transport::{smtp::client::Tls, smtp::client::TlsParameters, smtp::SmtpTransport},
@@ -6,15 +6,6 @@ use lettre::{
 };
 
 use crate::config::model::Account;
-
-error_chain! {
-    links {
-        Config(crate::config::model::Error, crate::config::model::ErrorKind);
-    }
-    foreign_links {
-        Smtp(lettre::transport::smtp::Error);
-    }
-}
 
 pub fn send(account: &Account, msg: &lettre::Message) -> Result<()> {
     let smtp_relay = if account.smtp_starttls() {